Move `is_clickhouse_enabled` from `posthog.ee` to `posthog.utils`
